Back to uint8_t

+++ src/board.hpp
@@ -11,8 +11,8 @@ class cBoard{
 		vec2D m_size;
 
 		// Containers storing board state
-		std::vector<int>  m_tiles;
-		std::vector<bool> m_passableTiles;
+		std::vector<uint8_t> m_tiles;
+		std::vector<bool>    m_passableTiles;
 
 		std::set<int> m_occupiedByUnit;
 		std::set<int> m_occupiedByEnemy;
